Refactor out jsonapi internals to utils/jsonapi
