+++ utils/jsonapi/helpers_test.go
@@ -1,8 +1,9 @@
-package action
+package jsonapi
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_checkIsPublicSuffix(t *testing.T) {
